Update but343.1.xml
Proofreading edits made to Homer and Demosthenes plates. I didn't find spelling or grammar errors in the Homer descriptions or tags, but found two in the Demosthenes descriptions and tags. I changed an incorrect adverb from "recline" to "reclining" and changed "a" to "the" in both the longer image description, and in the character descriptions of Archais and one of the sentries.